// src/libs/api/schema/core-services.d.ts
export interface components {
    schemas: {
// ...
        Contact: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            legal_type: components["schemas"]["ContactLegalType"];
            /** Tax Id */
            tax_id: string | null;
            image: components["schemas"]["ContactImageURL"];
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /**
             * Last Modified At
             * Format: date-time
             */
            last_modified_at: string;
            created_by: components["schemas"]["UserMin"];
            last_modified_by: components["schemas"]["UserMin"] | null;
            /** Deleted At */
            deleted_at: string | null;
            parent: components["schemas"]["ContactMin"] | null;
            /** Tags */
            tags: components["schemas"]["TagMin"][];
        };
// ...
        ContactMin: {
            /** Pk */
            pk: number;
            /** Name */
            name: string;
            legal_type: components["schemas"]["ContactLegalType"];
            image: components["schemas"]["ContactImageURL"];
            parent: components["schemas"]["ContactMin"] | null;
        };
// ...
        UserMin: {
            /** Pk */
            pk: number;
            /** First Name */
            first_name: string;
            /** Last Name */
            last_name: string;
            /**
             * Email
             * Format: email
             */
            email: string;
            image: components["schemas"]["UserImageURL"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}